+++ dockerfile
@@ -1,16 +1,3 @@
-# Copyright (c) 2017 Cisco and/or its affiliates.
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at:
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 FROM ubuntu:xenial
 
 RUN apt-get update && \
@@ -29,4 +16,4 @@ RUN mkdir /var/vpp
 
 RUN ulimit -c unlimited
 
-CMD ./.libs/icmp_responder
+CMD ./.libs/icmpr-epoll
